requirements.txt
flask
tensorflow
numpy
torch
torchvision
pillow
